1. Tracklist entity added 2. Downloader updated to work with AFNetworking 3.0

+++ vkmusic/VKMAudioNodeDownloader.h
@@ -15,7 +15,7 @@
 
 + (void)downloadNode:(VKMAudioNode *)node
                store:(NSString *)path
-            progress:(void (^) (float))progress
+            progress:(void (^) (double))progress
           completion:(void (^) (BOOL))completion;
 
 @end
